+++ packages/spotlight/Spottable/Spottable.js
@@ -166,10 +166,8 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 
 		constructor (props) {
 			super(props);
+			this.isFocused = false;
 			this.isHovered = false;
-			this.state = {
-				spotted: false
-			};
 		}
 
 		componentDidMount () {
@@ -178,8 +176,8 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		componentDidUpdate (prevProps) {
-			// if the component is spotted and became disabled,
-			if (this.state.spotted && (
+			// if the component is focused and became disabled,
+			if (this.isFocused && (
 				(!prevProps.disabled && this.props.disabled) ||
 				(!prevProps.spotlightDisabled && this.props.spotlightDisabled)
 			)) {
@@ -224,7 +222,7 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		componentWillUnmount () {
-			if (this.state.spotted) {
+			if (this.isFocused) {
 				forward('onSpotlightDisappear', null, this.props);
 			}
 			if (lastSelectTarget === this) {
@@ -299,7 +297,7 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 
 		handleBlur = (ev) => {
 			if (ev.currentTarget === ev.target) {
-				this.setState({spotted: false});
+				this.isFocused = false;
 			}
 
 			if (Spotlight.isMuted(ev.target)) {
@@ -311,7 +309,7 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 
 		handleFocus = (ev) => {
 			if (ev.currentTarget === ev.target) {
-				this.setState({spotted: true});
+				this.isFocused = true;
 			}
 
 			if (Spotlight.isMuted(ev.target)) {